Reverted back to commit 0adeedc54aacc8d63156c1c770eb2d063c873a6e
I found a way to hide the icons with the taskbar using DWM's Modkey+b hotkey.

By creating a command to launch thingmenu and a command to kill thingmenu, it
can be quickly killed and relaunched whenever the taskbar is minimized.

This way is a lot cleaner than merging both projects together. As a result,
the extra code for keyboard functionality is no longer needed and can be removed.

- changed height scaling to 1.4 to match DWM status bar height
- changed icon click colors to be more consistent, removed inverted colors